npcテーブルのスキーマを修正（digest_itemカラム） fixes #218
git-svn-id: http://svn.l1j.org/l1j/trunk@190 698199e4-35c4-443b-82c4-7f4b4ae73c13

+++ db/schema/mysql/npc.sql
@@ -42,7 +42,7 @@ CREATE TABLE `npc` (
   `agro_gfx_id1` int(10) NOT NULL DEFAULT '-1',
   `agro_gfx_id2` int(10) NOT NULL DEFAULT '-1',
   `pickup_item` tinyint(1) unsigned NOT NULL DEFAULT '0',
-  `digest_item` int(1) unsigned NOT NULL DEFAULT '0',
+  `digest_item` tinyint(3) unsigned NOT NULL DEFAULT '0',
   `brave_speed` tinyint(1) unsigned NOT NULL DEFAULT '0',
   `hpr_interval` int(10) unsigned NOT NULL DEFAULT '0',
   `hpr` int(10) unsigned NOT NULL DEFAULT '0',
